+++ bark/models/behavior/motion_primitives/macro_actions.cpp
@@ -32,9 +32,9 @@ BehaviorMotionPrimitives::MotionIdx BehaviorMPMacroActions::AddMotionPrimitive(
 }
 
 Trajectory BehaviorMPMacroActions::Plan(
-    float delta_time, const world::ObservedWorld& observed_world) {
+    double delta_time, const world::ObservedWorld& observed_world) {
   SetBehaviorStatus(BehaviorStatus::VALID);
-  const float dt = integration_time_delta_;
+  const double dt = integration_time_delta_;
   const int num_trajectory_points =
       static_cast<int>(std::ceil(delta_time / dt)) + 1;
